Add files via upload
The primary goal is to simplify the administrative tasks associated with employee attendance and vacation management. This requires creating a comprehensive system capable of efficiently handling the various aspects of these procedures, the target group of the system is employees. By allowing managers to easily log employee attendance, upload and view attendance records, manage leave requests instantly (including the ability to approve or deny requests), Maintaining a record of vacations within the system and extracting it when needed.